completed

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
  */
 void print_number(int n)
 {
-	long m;
+	int m;
 	int c;
 	long num;
